+++ code/server/globalv2loader.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "../qcommon/q_shared.h"
-qint
-xglobal_load_c(const qchar *path);
-unsigned qchar
-xglobal_flags(qchar *ip);